renamed profiles to startups
