+++ BackEnd/profiles_service/models/regions.py
@@ -1,6 +1,6 @@
 from sqlalchemy.orm import Mapped, mapped_column, relationship
 from models import Model
-from models.profiles import StartupProfileOrm
+from models.startups import StartupProfileOrm
 from sqlalchemy import ForeignKey
